saner code
