+++ src/plugins/reviewDB/entities.ts
@@ -1,6 +1,6 @@
 /*
  * Vencord, a modification for Discord's desktop app
- * Copyright (c) 2022 Vendicated and contributors
+ * Copyright (c) 2023 Vendicated and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,7 +16,11 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-import { Badge } from "./Badge";
+export const enum UserType {
+    Banned = -1,
+    Normal = 0,
+    Admin = 1
+}
 
 export const enum ReviewType {
     User = 0,
@@ -25,7 +29,36 @@ export const enum ReviewType {
     System = 3
 }
 
-export interface Sender {
+export interface Badge {
+    name: string;
+    description: string;
+    icon: string;
+    redirectURL: string;
+    type: number;
+}
+
+export interface BanInfo {
+    id: string;
+    discordID: string;
+    reviewID: number;
+    reviewContent: string;
+    banEndDate: number;
+}
+
+export interface ReviewDBUser {
+    ID: number;
+    discordID: string;
+    username: string;
+    profilePhoto: string;
+    clientMod: string;
+    warningCount: number;
+    badges: any[];
+    banInfo: BanInfo | null;
+    lastReviewID: number;
+    type: UserType;
+}
+
+export interface ReviewAuthor {
     id: number,
     discordID: string,
     username: string,
@@ -37,7 +70,7 @@ export interface Review {
     comment: string,
     id: number,
     star: number,
-    sender: Sender,
+    sender: ReviewAuthor,
     timestamp: number;
     type?: ReviewType;
 }
